Backport rpi_panel_v2 regulator
regulator: Add a regulator for the new LCD panels

The newer Raspberry Pi 5" and 7" panels have a slightly different
register map to the original one.
Add a new driver for this regulator.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

regulator: rpi_panel_v2: Add remove and shutdown hooks

Add shutdown and remove hooks so that the panel gets powered
off with the system.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

regulator: rpi_panel_v2: Add delay on I2C reads

As with many microcontrollers, the panel wants to use clock
stretching during a read so that the appropriate return value
can be generated and programmed into the hardware.
With Pi0-3 really not supporting clock stretching, this resulted
in the panel firmware dying.

Insert a delay between the write and the read to give the firmware
a chance to generate the relevant return value.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

drivers: media: pci: Add wrapper after removal of follow_pfn